build: add option to create JSON info files
The JSON info files contain details about the created firmware images
per device and are stored in $(TOPDIR)/json/. Finally all JSON files are
merged in a devices.json. This allows tools to parse the devices.json
file and once the desired device is found based on device title,
advanced information are retrived via the <device_id>.json file.

An example of devices.json:

    {
        "release_version": "",
        "release_commit": "r10374+2-fdd0a8d491",
        "devices": {
            "GL.iNet GL-AR150": "glinet_gl-ar150"
            ,"GL.iNet GL-AR300M-Lite": "glinet_gl-ar300m-lite"
            ,"GL.iNet GL-AR300M": "glinet_gl-ar300m-nor"
            ,"GL.iNet GL-AR750S": "glinet_gl-ar750s"
            ,"GL.iNet GL-X750": "glinet_gl-x750"
        }
    }

An example of glinet_gl-ar150.json

    {
        "id": "glinet_gl-ar150",
        "title": "GL.iNet GL-AR150",
        "vendor": "",
        "model": "",
        "variant": "",
        "supported_devices": [ "glinet,gl-ar150", "gl-ar150" ],
        "target": "ath79",
        "subtarget": "generic",
        "images": [
        {
            "type": "sysupgrade", "name": "openwrt-ath79-generic-glinet_gl-ar150-squashfs-sysupgrade.bin", "hash": "1e1767ffa0d1ff223f5d8a69ef7a87765e6a9ec62bc736a0f7a598a4cd4b3efa" }
        ]
    }

This could greatly simplify tooling for image retreival in future.

The option of JSON creatin is disabled by default, but active for
buildbots.

Signed-off-by: Paul Spooren <mail@aparcar.org>