build: store used build PROFILE on device
Currently the build PROFILE and `board_name` are not necessary the same
(e.g. x86/64, mvebu/cortexa9) which makes it hard to know what image to
install for an upgrade.

This patch stores the used profile in `/etc/os-release` which allows it
to be used by scripts.

The PROFILE_SANITIZED variable was moved from `image.mk` to `version.mk`
to be available for the `VERSION_SED_SCRIPT` command. As `version.mk` is
imported by `image.mk` anyway this move appears to be safe.

Signed-off-by: Paul Spooren <mail@aparcar.org>