mvebu,cortexa9: rename linksys,rango to wrt3200acm
This patch was previously send to the Kernel, however as it also changes
labels and the DTS filename existing devices could break.

The internal Linksys codename `rango` is renamed to `wrt3200acm`, it's
common model name. This way the build code base becomes more readable as
developers don't need to keep track of two different names for the same
device.

Additionally running devices now know which profile was used to create
the running firmware, instead of requiring an additional mapping.

Signed-off-by: Paul Spooren <mail@aparcar.org>