squashfs4: Add package
The original `squashfs-tools`  was forked and renamed to `squashfskit`
to integrate a variation of distro patches. This happened in 889b6423b7.

A new release of the original `squashfs-tools` came out including such
patches, making the fork `squashfskit` obsolete.

This commits adds the packages again including the latest patches
applied to `squashfskit`. For clarification the package is called
`squashfs4` instead of `squashfs-tools`, as there is already a legacy
version called `squashfs3`.

Signed-off-by: Paul Spooren <mail@aparcar.org>